photo_gallery/requirements.txt
django==3.2.4
djangorestframework==3.12.4
python-decouple==3.4
drf-yasg==1.20.0
Pillow==8.2.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
pytest==6.2.4
pytest-django==4.3.0
pytest-cov==2.12.1
model-bakery
pytest-mock==3.6.1





